feat: preprocessing functions moved into preprocesing.py #29
